# backend/api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8443
  description: Generated server url
paths:
...
  /api/books/{id}:
    get:
      tags:
      - book-api-controller
      operationId: getBookById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Book not found
        "200":
          description: BOOK FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Element'
        "400":
          description: Invalid id supplied
    put:
      tags:
      - book-api-controller
      operationId: updateBook
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ElementDTO'
        required: true
      responses:
        "200":
          description: BOOK PUT CORRECTLY
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Element'
        "404":
          description: Book not found
        "400":
          description: Invalid id supplied
    delete:
      tags:
      - book-api-controller
      operationId: deleteBook
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: BOOK DELETED CORRECTLY
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Element'
        "404":
          description: Book not found
        "401":
          description: Unauthorized
        "400":
          description: Invalid id supplied
  /api/books/{id}/image:
    get:
      tags:
      - book-api-controller
      operationId: getBookImageById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Book image not found
        "200":
          description: GET BOOK IMAGE CORRECTLY
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Element'
        "400":
          description: Invalid id supplied
    put:
      tags:
      - book-api-controller
      operationId: updateBookImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              required:
              - imageFile
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
      responses:
        "200":
          description: BOOK IMAGE PUT CORRECTLY
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Element'
        "404":
          description: Book not found
        "400":
          description: Invalid id supplied
    post:
      tags:
      - book-api-controller
      operationId: uploadBookImageById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              required:
              - imageFile
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
      responses:
        "404":
          description: Book not found
        "201":
          description: BOOK IMAGE POSTED CORRECTLY
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Element'
        "400":
          description: Invalid id supplied
    delete:
      tags:
      - book-api-controller
      operationId: deleteBookImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: BOOK IMAGE DELETED CORRECTLY
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Element'
        "404":
          description: Book not found
        "401":
          description: Unauthorized
        "400":
          description: Invalid id supplied
...
  /api/books/:
    get:
      tags:
      - book-api-controller
      operationId: getBooks
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageElement'
    post:
      tags:
      - book-api-controller
      operationId: uploadBook
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ElementDTO'
        required: true
      responses:
        "404":
          description: Book not found
        "201":
          description: BOOK POSTED CORRECTLY
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Element'
        "400":
          description: Invalid id supplied
...
  /api/elements/states:
    get:
      tags:
      - element-api-controller
      operationId: getState
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: string
                  enum:
                  - COMPLETO
                  - EN_EMISION
                  - POR_ESTRENARSE
  /api/elements/season:
    get:
      tags:
      - element-api-controller
      operationId: getSeason
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: string
                  enum:
                  - OTOÑO
                  - INVIERNO
                  - PRIMAVERA
                  - VERANO
  /api/elements/genres:
    get:
      tags:
      - element-api-controller
      operationId: getGenres
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: string
                  enum:
                  - ACCION
                  - TERROR
                  - AVENTURA
                  - MISTERIO
                  - ROMANCE
                  - CIENCIAFICCION
                  - DRAMA
                  - INFANTIL
                  - COMEDIA
                  - FANTASIA
                  - SOBRENATURAL
                  - NOVELA
                  - JUVENIL
  /api/books/top:
    get:
      tags:
      - book-api-controller
      operationId: getTop5Books
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageElement'
components:
  schemas:
...
    Element:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        author:
          type: string
        year:
          type: integer
          format: int32
        type:
          type: string
        season:
          type: string
        state:
          type: string
        country:
          type: string
        base64Image:
          type: string
        generos:
          type: array
          items:
            type: string
    ElementDTO:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        author:
          type: string
        year:
          type: integer
          format: int32
        type:
          type: string
        season:
          type: string
        state:
          type: string
        country:
          type: string
        genres:
          type: array
          items:
            type: string
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        unpaged:
          type: boolean
        paged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        unsorted:
          type: boolean
        sorted:
          type: boolean
...
    PageElement:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Element'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean